se unifoco los modulos de recargos, becas y descuentos en uno para mejor manejo y mantenimiento

+++ src/system/system-extra-charges/entities/system-extra-charges.entity.ts
@@ -36,13 +36,13 @@ export class SystemExtraCharges {
   })
   percentage: number;
 
-  @Column('tinyint', {
+  @Column('int', {
     nullable: true,
     name: 'dia_inicio',
   })
   startDay: number | null;
 
-  @Column('tinyint', {
+  @Column('int', {
     nullable: true,
     name: 'dia_fin',
   })
